requirements.txt
# ============================================
# AI Tutor - Python Dependencies
# ============================================
# 
# This file contains all Python packages required to run the AI Tutor
# multi-agent learning system. Dependencies are organized by category
# for better understanding and maintenance.
# 
# Author: AI Tutor Team
# Version: 1.0.0
# Python Version: 3.10+
# 
# Installation: pip install -r requirements.txt
# ============================================

aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
curl_cffi==0.11.1
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.16
fastapi==0.115.12
filelock==3.18.0
Flask==3.0.0
Flask-Cors==4.0.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.5.1
google-adk==0.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.94.0
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.16.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-aiohttp==0.1.4
httpx-sse==0.4.0
huggingface-hub==0.32.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
litellm==1.71.1
MarkupSafe==3.0.2
mcp==1.9.1
multidict==6.4.4
multitasking==0.0.11
numpy==2.2.6
openai==1.82.0
opentelemetry-api==1.33.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==25.0
pandas==2.2.3
peewee==3.18.1
platformdirs==4.3.8
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psutils==3.3.10
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyparsing==3.2.3
pypdf==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.5
starlette==0.46.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
yfinance==0.2.61
zipp==3.22.0
